Fix Use Cases

+++ src/main/java/com/ari/wishlist/application/usecase/DeleteWishlistUseCase.java
@@ -1,5 +1,6 @@
 package com.ari.wishlist.application.usecase;
 
+import com.ari.wishlist.domain.exception.WishlistNotFoundException;
 import com.ari.wishlist.domain.repository.WishlistRepository;
 import org.springframework.stereotype.Component;
 
@@ -12,6 +13,10 @@ public class DeleteWishlistUseCase {
     }
 
     public void execute(String customerId) {
+        boolean exists = wishlistRepository.existsByCustomerId(customerId);
+        if (!exists) {
+            throw new WishlistNotFoundException("Wishlist not found for customer ID: " + customerId);
+        }
         wishlistRepository.deleteByCustomerId(customerId);
     }
 }
